pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "keyboard-shopping-agent"
version = "0.1.0"
description = "Personalized keyboard shopping pipeline — 10x better than Google Shopping, no sponsored links."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "keyboard-shopping-agent contributors"},
]
keywords = ["keyboard", "ergonomic", "shopping", "comparison", "mechanical-keyboard"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business",
]

dependencies = [
    "httpx>=0.27,<1",
    "openpyxl>=3.1,<4",
    "beautifulsoup4>=4.12,<5",
    "lxml>=5.0,<6",
    "rich>=13.0,<14",
]

[project.optional-dependencies]
google = [
    "google-auth>=2.29,<3",
    "google-auth-oauthlib>=1.2,<2",
    "google-api-python-client>=2.120,<3",
]
dev = [
    "pytest>=8.0,<9",
    "pytest-cov>=5.0,<6",
    "ruff>=0.5,<1",
]
all = [
    "keyboard-shopping-agent[google,dev]",
]

[project.scripts]
ksa = "scripts.run_pipeline:main"

[tool.setuptools.packages.find]
include = ["src*", "scripts*"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]

[tool.ruff.lint.per-file-ignores]
"src/adapters/amazon_adapter.py" = ["E501"]
"src/adapters/bestbuy_adapter.py" = ["E501"]
"src/adapters/walmart_adapter.py" = ["E501"]
"src/enrichment/reviews.py" = ["E501"]
"src/output.py" = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
